Reorganize and simplify the invoice form code
